refactor: MainPage Body
