+++ client/src/Component/Molecules/index.tsx
@@ -1,2 +1,4 @@
 export * from "./MainPage/Main.Header.RightSide";
 export * from "./MainPage/Main.Header.Center";
+export * from "./MainPage/Main.Body.Left.ButtonContainer";
+export * from "./MainPage/Main.Body.Right.ButtonContainer";
