+++ client/src/Component/Organism/MainPage/Main.Body.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import MainBodyLeft from "./Main.Body.Left";
-import MainBodyRight from "./Main.Body.Right";
+import { MainBodyLeft } from "./Main.Body.Left";
+import { MainBodyRight } from "./Main.Body.Right";
 
 export const MainPageBody: React.FC = () => {
   return (
